requirements.txt
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-text-splitters>=0.0.1
sentence-transformers>=2.0.0
beautifulsoup4>=4.12.0
crawl4ai>=0.2.0
strip-markdown>=0.1.4
googlesearch-python>=1.2.3
duckduckgo-search>=4.0.0
